sync snapshots for rsync replication
Syncs ZFS snapshot between the hosts involved in the rsync
replication

Every rsync replication moves data from a snapshot, not the live
filesystem. Create a temporary snapshot specific to the replication id
on the source. If a temporary snapshot already exists, no new snapshot
is created.

The rsync shell script which runs the actual replication does the
snapshot sync post replication run. On a successful run, the
temporary snapshot is renamed on the source side. So, if there's a
temporary snapshot present, it's safe to assume that the previous
run for that replication id did not complete successfully.

Signed-off-by: six-k <ramsri.hp@gmail.com>